pyproject.toml
[project]
name = "autogen-core1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autogen-agentchat>=0.7.1",
    "autogen-core>=0.7.1",
    "autogen-ext[grpc]>=0.7.1",
    "ipykernel>=6.30.0",
    "openai>=1.97.1",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.9.0",
]
